// src/types.ts
export interface ICreateContractTemplateDTO {
  title: string
  type: string
  filePath: string
}

export interface IUpdateContractTemplateDTO {
  title?: string
  type?: string
  filePath?: string
}